+++ tests/php-mode-test.el
@@ -319,12 +319,9 @@ style from Drupal."
   "All static method should appear on imenu whether 'static' keyword is placed before or after visibility"
   (with-php-mode-test ("issue-83.php")
     (let* ((index-alist (imenu--make-index-alist))
-           (public-methods (mapcar 'car (cdr (assoc "Public Methods" index-alist))))
-           (all-methods (mapcar 'car (cdr (assoc "All Methods" index-alist)))))
-      (should (member "staticBeforeVisibility" public-methods))
-      (should (member "staticBeforeVisibility" all-methods))
-      (should (member "staticAfterVisibility" public-methods))
-      (should (member "staticAfterVisibility" all-methods)))))
+           (all-methods (mapcar 'car (cdr (assoc "Methods" index-alist)))))
+      (should (member "static public function staticBeforeVisibility()" all-methods))
+      (should (member "public static function staticAfterVisibility()" all-methods)))))
 
 (ert-deftest php-mode-test-issue-99 ()
   "Proper indentation for 'foreach' statements without braces."